change new API to use unsigned where it makes sense
darcs-hash:20100405054136-c8de0-a106e736e68aa365986a1cc91b9faf8492dd6ea1.gz

+++ api/general.c
@@ -96,7 +96,7 @@ static const char nlopt_algorithm_names[NLOPT_NUM_ALGORITHMS][256] = {
 
 const char *nlopt_algorithm_name(nlopt_algorithm a)
 {
-     if (a < 0 || a >= NLOPT_NUM_ALGORITHMS) return "UNKNOWN";
+     if (((int) a) < 0 || a >= NLOPT_NUM_ALGORITHMS) return "UNKNOWN";
      return nlopt_algorithm_names[a];
 }
